app/src/test/java/com/example/aoc24/DaySolution19Test.kt
package com.example.aoc24

import com.example.solutions.DaySolution19
import kotlinx.coroutines.test.runTest
import org.junit.Assert.assertEquals
import org.junit.Test

class DaySolution19Test {
    private val solution = DaySolution19(TestLogger())

    @Test
    fun testPart1() = runTest {
        input19.forEachIndexed { id, str -> solution.part1.handleLine(str, id) }
        solution.part1.finish()
        assertEquals(result, solution.part1.obtainResult())
    }

    @Test
    fun testPart2() = runTest {
        input19.forEachIndexed { id, str -> solution.part2.handleLine(str, id) }
        solution.part2.finish()
        assertEquals(result, solution.part2.obtainResult())
    }

    private val input1 = """r, wr, b, g, bwu, rb, gb, br

brwrr
bggr
gbbr
rrbgbr
ubwu
bwurrg
brgr
bbrgwb""".trimMargin().split("\n")



    private val result = "11"
}